requirements.txt
torch
torchvision
gradio
Pillow
numpy
